LARGE UPDATE: add an app/teams page and a form.
Update app\routes.py to include a sample dummy competition, which it
renders with the teams.html page, and a school form. It flashes the
values of these fields on submit.
Update base.html to add aesthetics - just a space under the navbar
app\forms.py contains classes that outline each form. SchoolForm is used
to create a new school with teams.
app\templates\teams.html is the template for the teams page and
currently just contains the form.